requirements.txt
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.28.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msrest==0.7.1
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
oauthlib==3.2.2
protobuf==3.20.3
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Unidecode==1.3.6
urllib3==2.0.3
Werkzeug==2.3.6
